Addition of two new features: (1) properties file (2) meta file.
1. Properties file
A new command line option (-p, --properties) has been added, accepting
the path to a properties file that contains the analysis options to be used.
If this option is given, all other options are ignored.

The skeleton of a properties file has been added, too. The skeleton is
created by invoking the createProperties gradle task.

The support of additional properties files (in $DOOP_HOME or $HOME) has
been droped.

2. Meta file
Upon startup, the analysis options are dumped to a "meta" file placed
in the output dir of the analysis.