Addition of artifact (jar) publication capability.
To cope with dependencies among Doop components, the doop-nexgen project can publish its jar in an ivy repository (ready to be utilized by the other doop projects such as doop-client and doop-gradle-plugin).

For the time being, the ivy repository is just a local directory which is a sibling of doop-nexgen. This feature will permit Anastasis and me to work on the tasks ahead, without affecting anyone else.

Once the dependent projects mature, we can decide on the details of a more adequate internal artifact publication architecture (e.g. host an internal plast maven repo on centauri or whatever suits our needs best).